requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
transformers==4.36.2
torch>=2.6.0
tqdm==4.66.1
python-dotenv==1.0.0
